Spec: Handle Ruby 1.8 <-> 1.9 String-encoding differences
NOTE: This removes one lambda that ensured that the test fails if
Ascii85.decode raises an exception. However, minitest does not appear to
have a "wont_raise" or similar assertion, so unhandled exceptions are
regarded as errors, and the testing process is aborted.

Not as clean as before, unfortunately, but not a big problem.