README: Remove `Contributors` section
I feel bad about removing this, but the truth is that I haven't been properly
updating this section of the README for the more recent contributions, and
having just the oldest contribution here doesn't feel right either.

Since I last released a version of this Gem, GitHub has also added its own
section to the repo UI that lists all contributors and their commits, which will
always be up-to-date, so having a manually curated section in the README is a
little bit redundant.