decode: Don't return the (frozen) ''-literal for the empty String
Instead, we `#dup` the literal to unfreeze it. This makes sure that there is no
case where the user is given a frozen String. This is a small foot-gun in so
far that the String returned by the happy path is not frozen, and people could
potentially trip over the frozen String returned when the input is invalid.

Ideally I'd want the returned String to always be frozen; however, that might
break existing code, so we just make sure that the returned String is always
thawed.