README: Add Limitations section
I tried to make an implementation that streams the input for encoding/decoding
using a lazy enumerator, but that proved to be too complicated for my tastes.

Since, in the 15-year lifespan of the gem, nobody complained about this problem,
I'll just assume that everyone is fine with only using it on small inputs; it
would be a bit dishonest to not add a disclaimer about this though, hence this
commit.